remove parallel flag and move test
